+++ src/mathjax/loadMathJax.js
@@ -2,7 +2,7 @@
 import load from './load'
 
 // mathjax cdn shutdown the 30/04/2017!!! https://cdn.mathjax.org/mathjax/latest/MathJax.js
-const DEFAULT_SCRIPT = 'https://cdnjs.cloudflare.com/ajax/libs/mathjax/2.7.0/MathJax.js'
+const DEFAULT_SCRIPT = 'https://cdnjs.cloudflare.com/ajax/libs/mathjax/2.7.5/MathJax.js'
 
 const DEFAULT_OPTIONS = {
   jax: ['input/TeX', 'output/CommonHTML'],